refactor: Route Handler params 처리, 토큰 추출 로직 헬퍼 함수 분리 (SM-40)
