+++ src/lib/tokenUtils.ts
@@ -8,30 +8,37 @@ interface ExtractTokensResult {
   stripped: unknown;
 }
 
+// 내부 헬퍼 함수
+const findTokenInObject = (obj: Record<string, unknown>, key: 'accessToken' | 'refreshToken'): string | null => {
+  // Top-level에서 찾기
+  if (typeof obj[key] === 'string') {
+    return obj[key] as string;
+  }
+
+  // data 객체 내부에서 찾기
+  if (obj.data && typeof obj.data === 'object') {
+    const data = obj.data as Record<string, unknown>;
+    if (typeof data[key] === 'string') {
+      return data[key] as string;
+    }
+  }
+
+  return null;
+};
+
 export const extractTokens = (json: unknown): ExtractTokensResult => {
-  if (!json || typeof json !== 'object') return { accessToken: null, refreshToken: null, stripped: json };
+  if (!json || typeof json !== 'object') {
+    return { accessToken: null, refreshToken: null, stripped: json };
+  }
 
   const obj = json as Record<string, unknown>;
 
-  const accessToken =
-    typeof obj.accessToken === 'string'
-      ? obj.accessToken
-      : obj.data &&
-          typeof obj.data === 'object' &&
-          typeof (obj.data as Record<string, unknown>).accessToken === 'string'
-        ? ((obj.data as Record<string, unknown>).accessToken as string)
-        : null;
+  const accessToken = findTokenInObject(obj, 'accessToken');
+  const refreshToken = findTokenInObject(obj, 'refreshToken');
 
-  const refreshToken =
-    typeof obj.refreshToken === 'string'
-      ? obj.refreshToken
-      : obj.data &&
-          typeof obj.data === 'object' &&
-          typeof (obj.data as Record<string, unknown>).refreshToken === 'string'
-        ? ((obj.data as Record<string, unknown>).refreshToken as string)
-        : null;
-
-  if (!accessToken && !refreshToken) return { accessToken: null, refreshToken: null, stripped: json };
+  if (!accessToken && !refreshToken) {
+    return { accessToken: null, refreshToken: null, stripped: json };
+  }
 
   const strippedTop = { ...obj };
   delete strippedTop.accessToken;
